make auth-return a component
